MPI init variadic function last argument replaced NULL with (mpi_ctx *)NULL
The C Standard allows NULL to be either an integer constant
or a pointer constant. While passing NULL as an argument to
a function with a fixed number of arguments will cause NULL
to be cast to the appropriate pointer type, when it is passed
as a variadic argument, this will not happen if sizeof(NULL) != sizeof(void *).
This is possible for several reasons:
- Pointers and ints may have different sizes on a platform where NULL
  is an integer constant
- The platform may have different pointer types with different sizes
  on a platform. In that case, if NULL is a void pointer, it is the
  same size as a pointer to char, which might be sized differently
  than the required pointer type.